stories：update：redux connect
