+++ src/components/elements/Select/index.stories.tsx
@@ -1,5 +1,8 @@
+import store from '@/stores';
 import { action } from '@storybook/addon-actions';
 import type { Meta, StoryObj } from '@storybook/react';
+import type { ReactNode } from 'react';
+import { Provider } from 'react-redux';
 import { Select } from '.';
 
 const meta = {
@@ -17,4 +20,10 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const NormalSelect: Story = {};
+const MockStore = ({ children }: { children: ReactNode }) => {
+  return <Provider store={store}>{children}</Provider>;
+};
+
+export const NormalSelect: Story = {
+  decorators: [(story: () => ReactNode) => <MockStore>{story()}</MockStore>],
+};
