+++ src/components/molecules/Header/index.stories.tsx
@@ -1,0 +1,25 @@
+import store from '@/stores';
+import type { Meta, StoryObj } from '@storybook/react';
+import type { ReactNode } from 'react';
+import { Provider } from 'react-redux';
+import { Header } from '.';
+
+const meta = {
+  title: 'Molecules/Header',
+  component: Header,
+  parameters: {
+    layout: 'centered',
+  },
+  tags: ['autodocs'],
+} satisfies Meta<typeof Header>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+const MockStore = ({ children }: { children: ReactNode }) => {
+  return <Provider store={store}>{children}</Provider>;
+};
+
+export const NormalHeader: Story = {
+  decorators: [(story: () => ReactNode) => <MockStore>{story()}</MockStore>],
+};
