--- src/components/molecules/SideMenu/index.stories.ts
@@ -1,16 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { SideMenu } from '.';
-
-const meta = {
-  title: 'Molecules/SideMenu',
-  component: SideMenu,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-} satisfies Meta<typeof SideMenu>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const NormalSideMenu: Story = {};
