+++ src/main.tsx
@@ -1,8 +1,8 @@
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from '../src/stores';
 import App from './App.tsx';
 import './index.css';
+import store from './stores/index.ts';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <Provider store={store}>
